First commit
change int to float
deplicate scenes

+++ Assets/GE18_03/S_Main/Player/MoveScript.cs
@@ -3,7 +3,7 @@ using UnityEngine.InputSystem;
 
 public class MoveScript : MonoBehaviour
 {
-    public int rotationSpeed = 1;
+    public float rotationSpeed = 1;
     private float initSpeed;
     public float walkSpeed = 1;
     public float runSpeed = 2;
